Adaptations to EFX type management refactoring.
